requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
docker==7.0.0
Flask==3.0.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
py==1.11.0
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
retry==0.9.2
urllib3==2.1.0
Werkzeug==3.0.1
